refactor: hoist duplicated helpers into platform modules
Four helpers had drifted into 3-7 near-identical copies each. Each is now
defined once under server/lib/platform/:

- guards.ts       isRecord (was 7 copies, one spelled differently)
- path-safety.ts  isSafeManifestPath (was 3 byte-identical copies of a
                  path-traversal guard, one per ecosystem manifest parser —
                  a hardening fix had to land three times to take effect)
- concurrency.ts  mapWithConcurrency (was 3 copies; keeps the discovery
                  variant's fail-fast + empty-input + concurrency>=1
                  handling, so the gate paths gain those too)
- crypto-utils.ts sha256Hex (was 3 copies with 3 different signatures,
                  moved off stable-json.ts where a hash helper did not
                  belong; digests are compared across these layers)

server/lib/tar-parser.js keeps its own sha256Hex: it is inlined verbatim
into the rendered sandbox worker and must stay import-free.
