+++ server/lib/ecosystems/npm/staged-publishes-discovery.ts
@@ -1,4 +1,5 @@
 import { type AppDb } from "../../../db/client";
+import { mapWithConcurrency } from "../../platform/concurrency";
 import { recordScanEvent } from "../../../db/events";
 import { markNpmConnectionUsed, updateNpmConnectionValidation } from "../../../db/npm-connections";
 import {
@@ -373,35 +374,6 @@ function isStartedStagedPublishScan(
   scan: StartedStagedPublishScan | null,
 ): scan is StartedStagedPublishScan {
   return scan !== null;
-}
-
-async function mapWithConcurrency<T, U>(
-  items: readonly T[],
-  concurrency: number,
-  worker: (item: T) => Promise<U>,
-): Promise<U[]> {
-  if (!items.length) return [];
-  const results = new Map<number, U>();
-  let next = 0;
-  let firstError: unknown = null;
-  const workerCount = Math.min(Math.max(1, concurrency), items.length);
-  const runners = Array.from({ length: workerCount }, async () => {
-    for (;;) {
-      if (firstError) return;
-      const index = next;
-      next += 1;
-      if (index >= items.length) return;
-      try {
-        results.set(index, await worker(items[index]));
-      } catch (err) {
-        firstError ??= err;
-        return;
-      }
-    }
-  });
-  await Promise.all(runners);
-  if (firstError) throw firstError;
-  return items.map((_, index) => results.get(index)!);
 }
 
 /**
