+++ server/lib/ecosystems/vscode/manifest.ts
@@ -1,3 +1,5 @@
+import { isRecord } from "../../platform/guards";
+import { isSafeManifestPath } from "../../platform/path-safety";
 import type { FileRecord, PackageJsonSummary } from "../../review";
 import { safeJson } from "../../review/rules";
 import type { TarSuspiciousEntry } from "../../tar-parser.js";
@@ -279,15 +281,4 @@ function isSafeExtensionId(value: string): boolean {
   const parts = value.split(".");
   if (parts.length !== 2) return false;
   return SAFE_PUBLISHER_RE.test(parts[0]) && SAFE_EXTENSION_NAME_RE.test(parts[1]);
-}
-
-function isSafeManifestPath(path: string): boolean {
-  if (!path || path.length > 512 || path.includes("\0") || path.includes("\\")) return false;
-  if (path.startsWith("/") || path.startsWith("../") || path.includes("/../")) return false;
-  if (/^[A-Za-z]:/.test(path)) return false;
-  return path.split("/").every((part) => part && part !== "." && part !== "..");
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === "object" && value !== null && !Array.isArray(value);
 }
