+++ server/lib/workflow-gates/resolve.ts
@@ -1,3 +1,4 @@
+import { mapWithConcurrency } from "../platform/concurrency";
 import {
   type ResolvedReleaseBundle,
   type ResolvedReleaseFile,
@@ -110,23 +111,4 @@ export function compactDuplicateTextSamples(
     }
   }
   return artifact;
-}
-
-async function mapWithConcurrency<T, U>(
-  items: T[],
-  concurrency: number,
-  worker: (item: T) => Promise<U>,
-): Promise<U[]> {
-  const results = new Map<number, U>();
-  let next = 0;
-  const workers = Array.from({ length: Math.min(concurrency, items.length) }, async () => {
-    for (;;) {
-      const index = next;
-      next += 1;
-      if (index >= items.length) return;
-      results.set(index, await worker(items[index]));
-    }
-  });
-  await Promise.all(workers);
-  return items.map((_, index) => results.get(index)!);
 }
