+++ test/workers/scan-artifact-cleanup.test.ts
@@ -16,7 +16,8 @@ import {
   deleteScanArtifacts,
   writeScanArtifacts,
 } from "../../server/lib/scan/artifacts";
-import { sha256Hex, stableJson } from "../../server/lib/platform/stable-json";
+import { sha256Hex } from "../../server/lib/platform/crypto-utils";
+import { stableJson } from "../../server/lib/platform/stable-json";
 
 // Mirrors the private safeSegment in scan-artifacts.ts: object keys live under
 // `orgs/{seg(orgId)}/scans/{seg(scanId)}/...`, so listing/asserting in the test
